+++ extensions/lucide-icons/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Lucide Icons Changelog
 
+## [Update] - 2024-01-14
+
+- Removed stroke red color to mantain the icons as unchanged as possible
+- Updated to latest Lucide Icons version
+- Updated dependencies
+
 ## [Update] - 2023-12-27
 
 - Updated extension